+++ scripts/install-angie.sh
@@ -29,6 +29,7 @@ readonly CFM_LUA_MANIFEST=(
     cfm.lua cfm_panel.lua cfm_panel_tunnel.lua cfm_rules.lua cfm_stats.lua
     cfm_waf.lua cfm_waf_util.lua cfm_waf_detectors.lua cfm_waf_excl.lua
     cfm_clamav.lua cfm_cache_log.lua cfm_clearance.lua cfm_geo.lua cfm_purge.lua
+    cfm_filecache.lua cfm_origin_ka.lua cfm_bridge_cfg.lua
     log-cfm.lua sslcollector.lua
 )